Loads of Changes
I've more or less begun a major rework of this program. I am hoping to
take it to an MVC pattern, as well as add support for multiple Visuals,
all held in a convenient tabbed pane. I expect numerous other additions
in the future, but for now I'd like to focus on completing the base
refactoring.