Added (horrible) two-level listener system
After trying to fool around to get each TabView to have listeners that
reach the Controller directly, I resorted to using a button in the main
GUI to piggyback the info back to the controller. Pretty horrific. I
need to find a better way to send info, and will continue looking into a
more versatile listener. Years from now, I will be laughing at this
code.